+++ src-tauri/migrations/20240106020657_currency_to_travel.sql
@@ -1,0 +1,3 @@
+-- Add migration script here
+ALTER TABLE travel
+    ADD currency TEXT NOT NULL DEFAULT 'EUR';