+++ src-tauri/src/app/mod.rs
@@ -1,3 +1,3 @@
 pub mod travel;
-pub mod utils;
 pub mod country;
+pub mod currency;