+++ src-tauri/src/main.rs
@@ -15,8 +15,13 @@ async fn main() {
         .manage(conn)
         .invoke_handler(tauri::generate_handler![
             app::travel::get_travels,
+            app::travel::get_travel,
             app::travel::create_travel,
+            app::travel::update_travel,
+
             app::country::get_countries,
+
+            app::currency::get_currencies,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");