+++ src/enums.ts
@@ -1,0 +1,4 @@
+export enum TravelFormMode {
+    ADD = 'add',
+    EDIT = 'edit',
+}