+++ src/main.tsx
@@ -1,10 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import TravelManager from './components/TravelManager/TravelManager.tsx'
+import TravelManager from '@/components/Travel/TravelManager.tsx'
 import './index.css'
+import {Toaster} from "@/components/ui/toaster.tsx";
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <TravelManager />
-  </React.StrictMode>,
+    <React.StrictMode>
+        <TravelManager/>
+        <Toaster/>
+    </React.StrictMode>,
 )